time card 1.4

+++ categoryview.cpp
@@ -17,7 +17,7 @@ CategoryView::CategoryView(QWidget* parent)
   m_imagePicker = new JDatabasePicker(IMAGE_SQL_TABLE_NAME,
                                       tr("Imagem"),
                                       QIcon(":/icons/res/icon.png"),
-                                      JDatabasePicker::Flags::TextGroup);
+                                      (int)JDatabasePicker::Flags::TextGroup);
 
   QFormLayout* namelayout = new QFormLayout;
   namelayout->setContentsMargins(0, 0, 0, 0);
