+++ noteview.cpp
@@ -179,7 +179,7 @@ NoteView::NoteView(QWidget *parent)
   m_supplierPicker = new JDatabasePicker(PERSON_SQL_TABLE_NAME,
                                          tr("Fornecedor"),
                                          QIcon(":/icons/res/supplier.png"),
-                                         JDatabasePicker::Flags::TextPlaceholder);
+                                         (int)JDatabasePicker::Flags::TextPlaceholder);
 
   QVBoxLayout* vlayout1 = new QVBoxLayout();
   vlayout1->addLayout(hlayout2);
