+++ productbarcodeview.cpp
@@ -16,7 +16,7 @@ ProductBarcodeView::ProductBarcodeView(QWidget* parent)
   m_productPicker = new JDatabasePicker(PRODUCT_SQL_TABLE_NAME,
                                          tr("Produto"),
                                          QIcon(":/icons/res/item.png"),
-                                         JDatabasePicker::Flags::TextGroup);
+                                         (int)JDatabasePicker::Flags::TextGroup);
 
   m_edCode = new JLineEdit(JLineEdit::Input::All,
                            JLineEdit::st_defaultFlags1);
